# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: BANCO Asset Marketplace API — Cars, Real Estate, Industrial
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [INVALID_DATA, NOT_FOUND, UNAUTHORIZED, INTERNAL_ERROR, FORBIDDEN, RATE_LIMITED]
        message:
          type: string
    Meta:
      type: object
      properties:
        cursor:
          type: string
        has_next:
          type: boolean
        total:
          type: integer
...
    FeedItem:
      type: object
      description: >
        Immutable v1 contract — exactly 11 fields.
        No renames, no removals allowed in v1. Breaking changes go to /v2/*.
      required:
        - id
        - media_preview
        - price_display
        - title
        - location
        - trust_signal
        - has_video
        - is_sponsored
      properties:
        id:
          type: string
          description: Listing identifier (use for detail navigation).
        media_preview:
          type: string
        price_display:
          type: string
        installment_badge:
          type: string
          nullable: true
        title:
          type: string
        location:
          type: string
        urgency_signal:
          type: string
          nullable: true
        trust_signal:
          type: string
        smart_badge:
          type: string
          nullable: true
        has_video:
          type: boolean
        is_sponsored:
          type: boolean
        coordinates:
          $ref: "#/components/schemas/Coordinates"
          nullable: true
        best_offer_badge:
          type: string
          nullable: true
        industrial_type:
          type: string
          nullable: true
          description: Industrial sub-type (factory, warehouse, land, production_line, raw_material, machine) used for client-side category grouping. Null for non-industrial listings.
        category:
          type: string
          enum: [car, real_estate, industrial]
          nullable: true
          description: Additive. The listing's main section — lets clients adapt per-section UI (e.g. the save glyph) without re-deriving it.
        is_active:
          type: boolean
          nullable: true
          description: True when the listing status is "active" (publicly visible and promotable). Lets owner-facing surfaces built on FeedItem (e.g. the profile grid) gate the Promote control without fetching ListingDetail. Null when status is unknown.
        whatsapp_enabled:
          type: boolean
          nullable: true
          description: True when the seller opted this listing in to WhatsApp contact. Opt-in only (default false). Lets surfaces decide whether to surface a WhatsApp action. Null when unknown.
        created_at:
          type: string
          nullable: true
          description: Listing creation timestamp (ISO 8601). Owner-facing grids (profile, "my listings") render a "Listed <date>" caption. Null only when the source row lacks it; public feeds always carry it.
        is_request:
          type: boolean
          nullable: true
          description: True when this is a buyer "request/wanted" post (looking to buy) rather than a sale listing. Surfaces render a "طلب / Wanted" badge and the requests filter keys off it. Null when unknown.
        is_bookable:
          type: boolean
          nullable: true
          description: >-
            Additive. True when this listing is a furnished/daily rental (hotel
            model) that can be reserved from its detail. Lets the feed, search
            card, and map pin surface a "قابل للحجز / Bookable" affordance without
            fetching the detail. Null/false for everything else.
    Coordinates:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
...
paths:
...
  /v1/listings:
    get:
      operationId: getListings
      summary: Browse public listings
      tags: [listings]
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: category
          schema:
            type: string
            enum: [car, real_estate, industrial]
      responses:
        "200":
          description: Paginated listing feed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/FeedItem"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
    post:
      operationId: createListing
      summary: Create listing
      tags: [listings]
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, category, location, specs, media]
              properties:
                title:
                  type: string
                description:
                  type: string
                category:
                  type: string
                  enum: [car, real_estate, industrial]
                base_price_cash:
                  type: number
                  description: Required for normal sale listings. Optional when is_request is true (a buyer request).
                is_request:
                  type: boolean
                  description: When true, this is a buyer "request/wanted" post. Price (base_price_cash) becomes optional and a description is required.
                location:
                  type: string
                latitude:
                  type: number
                  description: Optional precise pin (send with longitude). Overrides the area centroid for near-me search + map display.
                longitude:
                  type: number
                  description: Optional precise pin (send with latitude).
                specs:
                  type: object
                  additionalProperties: true
                media:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [image, video]
                      url:
                        type: string
                      thumbnail_url:
                        type: string
                      is_thumbnail:
                        type: boolean
                payment_options:
                  type: array
                  items:
                    type: object
                    properties:
                      mode:
                        type: string
                        enum: [cash, seller_installment, bank_finance]
                      down_payment:
                        type: number
                      monthly_payment:
                        type: number
                      duration_months:
                        type: integer
                      is_islamic_compliant:
                        type: boolean
                logistics:
                  type: object
                  description: >
                    Additive (Task #40). Optional seller-provided logistics &
                    delivery block. All fields optional/nullable.
                  properties:
                    delivery_time_days:
                      type: integer
                      nullable: true
                    origin_type:
                      type: string
                      enum: [local, imported]
                      nullable: true
                    country_of_origin:
                      type: string
                      nullable: true
                    shipping_method:
                      type: string
                      enum: [container, bulk, air]
                      nullable: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...